requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30

SQLAlchemy>=2.0
asyncpg>=0.29
greenlet>=3.0
psycopg[binary]>=3.2
alembic>=1.13

pydantic-settings>=2.5
python-dotenv>=1.0
httpx>=0.27

google-auth>=2.30
python-jose[cryptography]>=3.3

pre-commit>=3.8

# Testing & Quality
pytest>=8.3
pytest-asyncio>=0.24
ruff>=0.6
mypy>=1.11
